requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-dotenv==1.1.1
PyJWT==2.8.0
pymongo==4.6.1
requests==2.31.0
python-socketio==5.10.0 